Updates to potential function.

+++ WellSolver.h
@@ -1,7 +1,7 @@
 #ifndef WELLSOLVER_H
 #define WELLSOLVER_H
 
-double potential(int a, int b, double x, double y);
+double potential(double a, double b, double x, double y);
 double wavefunction (double a, double b, double x, double y, int n);
 double x_length (double a, double b, double y);
 double y_length (double a, double b, double x);
